requirements.txt
# For Phase 1
requests>=2.31.0
pandas>=2.0.0
PyPDF2>=3.0.0
pathlib
numpy>=1.24.0
lxml>=4.9.0

# Phase 2
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Phase 3
google-generativeai>=0.3.0